Split opengl into 2 files
